+++ main.js
@@ -6,15 +6,13 @@ const fetch = require("node-fetch");
 class Application {
     constructor() {
         this.window = this.createWindow();
+        this.pip = null;
+
         this.initMainFolder();
         this.handleEvents();
     }
 
     handleEvents() {
-        ipcMain.on("window-minimize", () => this.window.minimize());
-        ipcMain.on("window-maximize", () => (this.window.isMaximized()) ? this.window.unmaximize() : this.window.maximize());
-        ipcMain.on("window-close", () => this.window.close());
-
         this.window.on("resize", () => {
             const { x, y, width, height } = this.window.getBounds();
             const f = this.window.isMaximized();
@@ -25,6 +23,17 @@ class Application {
             const { x, y, width, height } = this.window.getBounds();
             const f = this.window.isMaximized();
             this.window.webContents.send("window-update", { x, y, width, height, f });
+        });
+
+        this.window.on("close", () => {
+            if (this.pip) this.closePIP();
+        });
+
+        ipcMain.on("window-minimize", () => this.window.minimize());
+        ipcMain.on("window-maximize", () => (this.window.isMaximized()) ? this.window.unmaximize() : this.window.maximize());
+        ipcMain.on("window-close", () => {
+            this.window.close();
+            if (this.pip) this.closePIP();
         });
 
         ipcMain.on("get-main-folder", (e) => {
@@ -45,6 +54,11 @@ class Application {
                 this.window.webContents.send("song-control", data);
             };
             this.window.setThumbarButtons(this.thumbnailButtons);
+        });
+
+        ipcMain.on("toggle-pip-mode", (e, data) => {
+            if (!this.pip) this.openPIP();
+            else this.closePIP();
         });
     }
 
@@ -75,7 +89,6 @@ class Application {
                 contextIsolation: false,
             },
         });
-        window.setMenuBarVisibility(false);
         window.loadFile("src/index.html");
         if (jsonData) if (jsonData.window.f) window.maximize();
 
@@ -149,7 +162,6 @@ class Application {
             },
             lastSongID: 0,
             playbackRate: "1",
-            pitch: "1",
         };
         const strSaveData = JSON.stringify(saveData, null, 2);
         if (!fs.existsSync(settingsFile)) fs.writeFile(settingsFile, strSaveData, (err) => {
@@ -178,6 +190,43 @@ class Application {
         if (!fs.existsSync(statsFile)) fs.writeFile(statsFile, strStatsData, (err) => {
             if (err) console.error("ERROR HK-205 => Could not write stats.json:", err);
         });
+    }
+
+    openPIP() {
+        const { x, y, width, height } = this.window.getBounds();
+        this.pip = new BrowserWindow({
+            x: x + width - 250,
+            y: y + 100,
+            width: 200,
+            height: 200,
+            minWidth: 200,
+            minHeight: 200,
+            maxWidth: 500,
+            maxHeight: 500,
+            frame: false,
+            alwaysOnTop: true,
+            webPreferences: {
+                nodeIntegration: true,
+                contextIsolation: false,
+            },
+        });
+
+        this.pip.loadFile("src/pip.html");
+
+        this.pip.on("resize", () => {
+            const { width, height } = this.pip.getBounds();
+            const max = Math.max(width, height);
+            this.pip.setSize(max, max);
+        });
+
+        this.pip.on("maximize", (e) => {
+            e.preventDefault();
+        });
+    }
+
+    closePIP() {
+        this.pip.close();
+        this.pip = null;
     }
 };